+++ tests/index.ts
@@ -4,4 +4,4 @@ import 'reflect-metadata';
 
 // tests
 import './Angular2Apollo';
-import './ApolloQueryPipe';
+import './SelectPipe';